Fixed a board read issue
For some reason the board numbers were 1 higher in the row position. This meant each letter was one position off. To fix I had it run for 1 longer and just subtract 1 from each position. This is FINE because we already check to make sure the position isn't null, so even though it probably is hitting null on one of those it's not impacting the array.